Safety changes

+++ src/main/java/controller/WaterReportScreenController.java
@@ -557,7 +557,7 @@ public class WaterReportScreenController implements Initializable {
         });
         yAxis.setTickLabelFormatter(new StringConverter<Number>() {
             @Override public String toString(Number d) {
-                return (String.format("%.2f", d));
+                return (String.format("%.2f", (Double)d));
             }
             @Override public Number fromString(String d) {
                 return (Double.valueOf(d));
